ecosystem PM2

+++ fileUploader.js
@@ -11,24 +11,33 @@ if (!fs.existsSync(capFolder)) {
 }
 
 fs.readdir(capFolder, (err, files) => {
-    files.forEach(file => {
-        let filePath = path.join(capFolder, file);
-        fs.readFile(filePath, 'utf8', (err, data) => {
-            if (err) {
-                console.error(err)
-                return
-            }
-            const form = new FormData();
-            form.append('capture', fs.createReadStream(filePath));
-            axios.post(config.capturesEndpoint, form, { headers: form.getHeaders() })
-                .then(x => {
-                    unlink(filePath, (err) => {
-                        if (err) throw err;
-                        console.log(`Deleted ${filePath}`);
-                    })
-                }).catch(error =>
-                    console.log(JSON.stringify(error))
-                );
-        })
-    });
+    if (files.length > 0) {
+        console.log(`Found ${files.length + 1} files`);
+        files.forEach(file => {
+            let filePath = path.join(capFolder, file);
+            fs.readFile(filePath, 'utf8', (err, data) => {
+                console.log(`Sending ${filePath}`);
+                if (err) {
+                    console.error(err)
+                    return
+                }
+                const form = new FormData();
+                form.append('capture', fs.createReadStream(filePath));
+                axios.post(config.capturesEndpoint, form, { headers: form.getHeaders() })
+                    .then(x => {
+                        console.log(`Sended ${filePath}`);
+                        unlink(filePath, (err) => {
+                            if (err) throw err;
+                            console.log(`Deleted ${filePath}`);
+                        })
+                    }).catch(error =>
+                        console.log(JSON.stringify(error))
+                    );
+            })
+        });
+    } else {
+        console.log('No Files Found');
+        process.exit(0);
+    }
+
 });
